refactor : exception 계층화 및 불필요 클래스 제거
