--- src/main/java/wooteco/subway/exception/DuplicatedNameException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.exception;
-
-public class DuplicatedNameException extends RuntimeException {
-
-    public DuplicatedNameException() {
-        super("이미 존재하는 이름입니다.");
-    }
-}
