--- src/main/java/wooteco/subway/exception/SectionAdditionException.java
@@ -1,7 +1,0 @@
-package wooteco.subway.exception;
-
-public class SectionAdditionException extends RuntimeException {
-    public SectionAdditionException(String message) {
-        super(message);
-    }
-}
