--- src/main/java/wooteco/subway/exception/SectionDeleteException.java
@@ -1,7 +1,0 @@
-package wooteco.subway.exception;
-
-public class SectionDeleteException extends RuntimeException {
-    public SectionDeleteException(String message) {
-        super(message);
-    }
-}
