--- src/main/java/wooteco/subway/exception/WrongDistanceException.java
@@ -1,7 +1,0 @@
-package wooteco.subway.exception;
-
-public class WrongDistanceException extends RuntimeException {
-    public WrongDistanceException(String message) {
-        super(message);
-    }
-}
