+++ src/main/java/wooteco/subway/exception/sectionexception/SectionAdditionException.java
@@ -1,0 +1,9 @@
+package wooteco.subway.exception.sectionexception;
+
+import wooteco.subway.section.model.Section;
+
+public class SectionAdditionException extends SectionException {
+    public SectionAdditionException(String message) {
+        super(message);
+    }
+}
