+++ src/main/java/wooteco/subway/exception/sectionexception/SectionDeleteException.java
@@ -1,0 +1,7 @@
+package wooteco.subway.exception.sectionexception;
+
+public class SectionDeleteException extends SectionException {
+    public SectionDeleteException(String message) {
+        super(message);
+    }
+}
