+++ src/main/java/wooteco/subway/exception/sectionexception/SectionException.java
@@ -1,0 +1,9 @@
+package wooteco.subway.exception.sectionexception;
+
+import wooteco.subway.section.model.Section;
+
+public class SectionException extends RuntimeException{
+    public SectionException(String message) {
+        super(message);
+    }
+}
