+++ src/main/java/wooteco/subway/exception/sectionexception/WrongDistanceException.java
@@ -1,0 +1,7 @@
+package wooteco.subway.exception.sectionexception;
+
+public class WrongDistanceException extends SectionException {
+    public WrongDistanceException(String message) {
+        super(message);
+    }
+}
